requirements.txt
# FreqTrade3 - Dependências Principais
# Gerenciadas com máxima segurança e compatibilidade

# === CORE FREQTRADE ===
freqtrade>=3.0.0,<4.0.0

# === WEB FRAMEWORK ===
Flask>=2.3.0,<3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
Flask-Limiter>=3.0.0

# === ANÁLISE DE DADOS ===
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
talib-binary>=0.4.25

# === GRÁFICOS E VISUALIZAÇÃO ===
plotly>=5.15.0
dash>=2.13.0

# === BANCO DE DADOS ===
SQLAlchemy>=2.0.0
sqlite3  # Built-in with Python

# === SEGURANÇA ===
cryptography>=41.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0

# === COMUNICAÇÃO ===
requests>=2.31.0
httpx>=0.24.0
telegram-send>=0.34

# === MONITORAMENTO ===
prometheus-client>=0.17.0
psutil>=5.9.0

# === TESTES ===
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# === UTILITÁRIOS ===
python-dateutil>=2.8.0
click>=8.1.0
colorama>=0.4.6
rich>=13.5.0

# === DESENVOLVIMENTO ===
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# === Yfinance para dados reais ===
yfinance>=0.2.18
ccxt>=4.0.0

# === BACKTESTING AVANÇADO ===
pyfolio>=0.9.2
zipline>=1.4.1

# === MACHINE LEARNING ===
scikit-learn>=1.3.0
scipy>=1.11.0

# === NOTIFICAÇÕES ===
pushbullet.py>=0.12.0
slack-sdk>=3.21.0

# === WEBSOCKETS ===
python-socketio>=5.8.0
eventlet>=0.33.0

# === VALIDAÇÃO ===
pydantic>=2.0.0
jsonschema>=4.19.0

# === GESTÃO DE CONFIGURAÇÃO ===
configparser  # Built-in with Python
pyyaml>=6.0.0

# === CRON JOBS ===
APScheduler>=3.10.0

# === LOGGING ===
structlog>=23.0.0
colorlog>=6.7.0

# === CHAVES API ===
python-binance>=1.0.16

# === COMPRESSÃO ===
zlib  # Built-in with Python
gzip  # Built-in with Python

# === TEMPO ===
pytz>=2023.3

# === SERIALIZAÇÃO ===
orjson>=3.9.0

# === PERFOMANCE ===
cython>=3.0.0
numba>=0.57.0

# === DASHBOARD ===
dash-bootstrap-components>=1.4.0
dash-daq>=0.5.0

# === WEBSERVER ===
gevent>=23.0.0
eventlet>=0.33.0

# === DEPLOYMENT ===
gunicorn>=21.0.0
whitenoise>=6.5.0

# === ENVIRONMENT ===
python-decouple>=3.8
